requirements.txt
annotated-types==0.7.0
anyio==4.8.0
backoff==2.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
langfuse==2.59.4
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-telegram-bot==21.10
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.2
